+++ KeireCore/Source/Scripting/ManagedRuntimeBindings.cpp
@@ -3,13 +3,14 @@
 namespace Keire::Detail
 {
     ManagedRuntimeBindingsScope::ManagedRuntimeBindingsScope(IScriptRuntimeServices* services) noexcept
-        : m_Foundation(services), m_Input(services), m_Physics(services), m_World(services), m_Rendering(services),
-          m_Ui(services)
+        : m_Assets(services), m_Foundation(services), m_Input(services), m_Physics(services), m_World(services),
+          m_Rendering(services), m_Ui(services)
     {
     }
 
     void RegisterManagedRuntimeBindings(Coral::ManagedAssembly& assembly)
     {
+        RegisterManagedRuntimeAssets(assembly);
         RegisterManagedRuntimeFoundation(assembly);
         RegisterManagedRuntimeInput(assembly);
         RegisterManagedRuntimePhysics(assembly);